+++ config/initializers/locale.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 # Whitelist locales available for the application.
-I18n.available_locales = [:en, :es, :arr]
+I18n.available_locales = [:en, :es]