+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancements
 - **Design Changes**
   - Designs can now be translated into different languages
+  - Admins can now add Spanish translations to existing designs
 - **Gem Changes**
   - Updated to rails 5.2.0.rc2
   - Updated to devise 4.4.3